### 2021-05-21 #3305 @whedon
[PRE REVIEW]: argodata: An R interface to oceanographic data from the International Argo Program
**Submitting author:** @paleolimbot (<a href="http://orcid.org/0000-0002-9415-4582">Dewey Dunnington</a>)
**Repository:** <a href="https://github.com/ArgoCanada/argodata" target ="_blank">https://github.com/ArgoCanada/argodata</a>
**Version:** 0.0.0.9000
