### 2021-05-21 #3305 @kyleniemeyer
Hi @paleolimbot, thanks for your interest in JOSS. Because your submission appears to be an API client for accessing data, which may not meet our [requirements for research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), the editorial board is going to review your submission to see if it meets our scope for review. Any additional comments you have would be welcome.

@kthyng your feedback on this would be particularly welcome
