### 2021-05-21 #3305 @paleolimbot
Thanks @kyleniemeyer!

In defense of our submission, I'll note a that (1) Argo is not an API, it is a collection of files on an FTP server that are inaccessible to many researchers; and (2) JOSS published [a paper describing a Python wrapper of the exact same data set](https://joss.theoj.org/papers/10.21105/joss.02425) several months ago. I'll also note that we followed the format of the [JOSS paper describing bcdata](https://joss.theoj.org/papers/10.21105/joss.02927), a package which performs a similar task (make data accessible to be used in research that would otherwise not be because it is too hard to access).

We're also happy to reframe our submission to highlight the features of argodata that go beyond data access!
