### 2021-05-21 #3305 @kthyng
@paleolimbot, @kyleniemeyer As a high level comment, I edited the `argopy` paper that was mentioned as being previously published with JOSS and it was a solid JOSS submission. That package facilitates data access, but does so really well and also provides other functionality to make it possible to easily work with Argo data. For data like this, I think the data doesn't really get used without these excellent packages.

@paleolimbot I see that you mention argopy in the paper. Is your package analogous to that one but written in R?
