### 2021-05-21 #3305 @paleolimbot
It's very similar! There is a shared goal between argopy and argodata (make it easier to access a truly amazing data set) and overlap in functionality (search, download, combine, and load profiles). The bit that's unique to argodata is that it uses tables/data frames rather than multidimensional arrays. It can also load all variables from all file types.
