+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   articles/virtual-mooring: virtual-mooring.html
-last_built: 2021-05-21T12:59Z
+last_built: 2021-05-28T17:16Z
 urls:
   reference: https://argocanada.github.io/argodata/reference
   article: https://argocanada.github.io/argodata/articles
