### 2021-05-30 #3305 @paleolimbot
I forgot the suggested reviewer step noted in the initial comment - a few that we had in mind were:

- Michael Sumner (mdsumner): Familiar with Argo and experienced R user. We use the abstraction of NetCDF he invented in the ncmeta and tidync packages in our submission extensively!
- Sam Albers (boshek): Author of bcdata and tidyhydat, both excellent data access packages that integrate with tidyverse functions
- Kimberlee Baldry (KimBaldry): Argo expert and experienced R user
- Ben Tupper (btupper): Oceanographer and experienced R user

As a potential conflict of interest note, I have made minor contributions to packages by Sam Albers and Michael Sumner in the past, and Michael Sumner has done the same for several packages I author.
