### 2021-06-05 #3305 @Kevin-Mattheus-Moerman
@paleolimbot I've removed the `query-scope` label as the board has determined this may be in scope. I will therefore proceed to find an editor to handle this submission.
