### 2021-06-29 #3305 @KimBaldry
> @boshek and @KimBaldry — Are you interested in reviewing this submission for JOSS, the Journal of Open Source Software? This is an open review which takes place on github with issues. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration.

I have to disclose that I am working 0.8FTE outside of academia and completing my PhD part-time. The package is useful to my PhD work and I can test its functionality against my own code. However, I do have limited time so I can't promise to meet short deadlines. If the editor and authors are happy to proceed with this in mind, I am happy to review the code in my own time.

I look forward to seeing this published!
