### 2021-08-10 #3305 @KimBaldry
Hi @kthyng, I will get started this weekend! I only need to compile a PDF with the review criteria and using the review checklist, correct? Where do I post this PDF once complete?

Other potential reviewers who may be interested are Marin Cornec or Raphaelle Sauzede. Both use Argo and R.
