### 2021-08-10 #3305 @kthyng
Hi @KimBaldry! No need to make a PDF. The outline of the review should be in the github review issue that I can open when we have our second reviewer. As issues come up during review, please open issues in the software repo itself and link back to the review issue (keeps down the clutter but keeps everything connected). There is a reviewer guide available as well as a look at the reviewer checklist that you will use: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#

Thanks for the reviewer suggestions. It's summer here and I've gotten a bit behind. First I will try one more of the author's suggestions.

@mdsumner Are you interesting in reviewing this submission for the Journal of Open Source Software? We ask for reviews in the next 4-6 weeks, but are often an iterative process with the author so it's best to start earlier if possible. More information can be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#
