### 2021-08-10 #3305 @mdsumner
Thanks for consideration, and I'll think about it - but I do feel a *bit close* to this ... I don't have heaps of actual Argo experience, and I know the author and @KimBaldry quite well. this might be a case where seeking another reviewer is advisable? I'll put some thought into it.
