### 2021-08-16 #3305 @kthyng
@gmaze As the author of ArgoPy, which went through JOSS review, do you have any colleagues in your group/area who use R who might want to review this package? Thanks.
