### 2021-08-25 #3305 @catsch
Hi Kristen,

I am not very familiar with this type of review but I can try, what is the expected dead line? I have a very busy september.
Tell me how should I proceed ?
Best
Catherine

--
Catherine Schmechtig

0681827411

LEFE-CYBER Database
http://www.obs-vlfr.fr/proof

BGC Argo
http://biogeochemical-argo.org/

https://github.com/catsch

OSU Ecce Terra
Sorbonne Université - CNRS - UMS 3455
4, Place Jussieu
75 252 Paris Cedex 05
Tour 45 - couloir 45-55
2eme etage - 207

De: "Kristen Thyng" ***@***.***>
À: "openjournals/joss-reviews" ***@***.***>
Cc: "Catherine Schmechtig" ***@***.***>, "Mention" ***@***.***>
Envoyé: Mardi 17 Août 2021 16:04:55
Objet: Re: [openjournals/joss-reviews] [PRE REVIEW]: argodata: An R interface to oceanographic data from the International Argo Program (#3305)

Thanks [ https://github.com/gmaze | @gmaze ] !

[ https://github.com/CATSH | @CATSH ] Are you interested in reviewing this submission for JOSS, the Journal of Open Source Software? This is an open review which takes place on github with issues. You can learn more about the process here: [ https://joss.readthedocs.io/en/latest/reviewer_guidelines.html | https://joss.readthedocs.io/en/latest/reviewer_guidelines.html ] . I would hope you could start the review in the next few weeks with the intention of being through the typically iterative process in 4-6 weeks. Thanks for your consideration.
