### 2021-08-25 #3305 @kthyng
Hi @catsch! Thanks for your willingness to help. We know everyone is quite busy, especially during a global pandemic, so we are asking for 4-6 weeks total for review, though it's better to start earlier if possible because the review process tends to be iterative between the reviewers and author(s). This submission has been waiting in pre-review for quite awhile, but we would just be happy to have two solid reviewers so we can move forward! Let me know what you think.
