### 2021-08-25 #3305 @catsch
Ok Fine by me
I ll try to do my best. I can start next week
Catherine

--
Catherine Schmechtig

0681827411

LEFE-CYBER Database
http://www.obs-vlfr.fr/proof

BGC Argo
http://biogeochemical-argo.org/

https://github.com/catsch

OSU Ecce Terra
Sorbonne Université - CNRS - UMS 3455
4, Place Jussieu
75 252 Paris Cedex 05
Tour 45 - couloir 45-55
2eme etage - 207

De: "Kristen Thyng" ***@***.***>
À: "openjournals" ***@***.***>
Cc: "Catherine Schmechtig" ***@***.***>, "Mention" ***@***.***>
Envoyé: Mercredi 25 Août 2021 17:11:24
Objet: Re: [openjournals/joss-reviews] [PRE REVIEW]: argodata: An R interface to oceanographic data from the International Argo Program (#3305)

Hi [ https://github.com/catsch | @catsch ] ! Thanks for your willingness to help. We know everyone is quite busy, especially during a global pandemic, so we are asking for 4-6 weeks total for review, though it's better to start earlier if possible because the review process tends to be iterative between the reviewers and author(s). This submission has been waiting in pre-review for quite awhile, but we would just be happy to have two solid reviewers so we can move forward! Let me know what you think.
