### 2021-08-25 #3659 @whedon
[REVIEW]: argodata: An R interface to oceanographic data from the International Argo Program
**Submitting author:** @paleolimbot (<a href="http://orcid.org/0000-0002-9415-4582">Dewey Dunnington</a>)
**Repository:** <a href="https://github.com/ArgoCanada/argodata" target ="_blank">https://github.com/ArgoCanada/argodata</a>
**Version:** v0.1.0
