### 2021-10-05 3c7edd7
fix undefined variable
