### 2021-10-05 fb0d919
remove bogus 'helper' for filtering rectangles (#42)
