### 2021-10-05 c330e6c
add argo_filter_paramter() (#41)
