+++ man/argodata-package.Rd
@@ -28,7 +28,7 @@ Useful links:
 Other contributors:
 \itemize{
   \item Jaimie Harbin (\href{https://orcid.org/0000-0003-3774-3732}{ORCID}) [contributor]
-  \item Chris Gordon [contributor]
+  \item Christopher Gordon (\href{https://orcid.org/0000-0002-1756-422X}{ORCID}) [contributor]
   \item Dan Kelley (\href{https://orcid.org/0000-0001-7808-5911}{ORCID}) [contributor]
   \item Clark Richards (\href{https://orcid.org/0000-0002-7833-206X}{ORCID}) [contributor]
 }
