+++ .github/workflows/test-coverage.yaml
@@ -21,6 +21,8 @@ jobs:
         with:
           use-public-rspm: true
 
+      - uses: r-lib/actions/setup-pandoc@v1
+
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
           extra-packages: covr
