+++ .github/workflows/pkgdown.yaml
@@ -23,9 +23,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
-          extra-packages:
-            - pkgdown
-            - lubridate
+          extra-packages: pkgdown, lubridate
           needs: website
 
       - name: Cache Argo downloads
