### 2021-10-05 40278a3
move pandoc install to correct action
