### 2021-10-05 a36a8ff
correct mistake re: delayed and realtime data modes (#38)
