### 2021-10-05 c82a55e
add Argo DOI where it was missing; fix links to Argo data management site
