+++ inst/joss/paper.bib
@@ -1,8 +1,11 @@
 
-@misc{argo_argo_2020,
+@article{argo,
 	title = {Argo float data and metadata from {Global} {Data} {Assembly} {Centre} ({Argo} {GDAC})},
+	year = {2021},
+	author = {Argo},
 	copyright = {Creative Commons Attribution 4.0 International},
-	url = {https://www.seanoe.org/data/00311/42182/}
+	journal = {SEANOE},
+	doi = {10.17882/42182}
 }
 
 @misc{noauthor_comprehensive_2021,
