+++ README.md
@@ -12,10 +12,11 @@ coverage](https://codecov.io/gh/ArgoCanada/argodata/branch/master/graph/badge.sv
 <!-- badges: end -->
 
 The goal of argodata is to provide a data frame-based interface to data
-generated by the [Argo floats program](https://argo.ucsd.edu). This
-package provides low-level access to all variables contained within Argo
-NetCDF files; for a higher-level interface with built-in visualization
-and quality control, see the [argoFloats
+generated by the [Argo floats program](https://argo.ucsd.edu)
+([doi:10.17882/42182](https://doi.org/10.17882/42182)). This package
+provides low-level access to all variables contained within Argo NetCDF
+files; for a higher-level interface with built-in visualization and
+quality control, see the [argoFloats
 package](https://github.com/ArgoCanada/argoFloats).
 
 ## Installation
@@ -51,19 +52,17 @@ occasionally; update these using `argo_update_data()`).
 
 The main workflow supported by argodata is:
 
-  - Start with `argo_global_prof()`, `argo_global_traj()`,
+-   Start with `argo_global_prof()`, `argo_global_traj()`,
     `argo_global_meta()`, `argo_global_bio_prof()`,
     `argo_global_bio_prof()`, or `argo_global_synthetic_prof()`. These
     functions return data frames that contain meta information about
     each data file in Argo.
-  - Use `argo_filter_radius()`, other `argo_filter_*()` functions, or
+-   Use `argo_filter_radius()`, other `argo_filter_*()` functions, or
     `dplyr::filter()` to subset the index using search criteria.
-  - Use extractor functions like `argo_prof_levels()`,
+-   Use extractor functions like `argo_prof_levels()`,
     `argo_traj_measurement()`, `argo_meta_config_param()`, and
     `argo_tech_tech_param()` to read information for all files in the
     index subset.
-
-<!-- end list -->
 
 ``` r
 library(tidyverse)
@@ -80,8 +79,8 @@ prof_lab_may_2020 <- argo_global_prof() %>%
 # download, cache, and load NetCDF files
 levels_lab_may_2020 <- prof_lab_may_2020 %>% 
   argo_prof_levels()
+#> Downloading 55 files from 'https://data-argo.ifremer.fr'
 #> Extracting from 55 files
-
 # plot!
 levels_lab_may_2020 %>% 
   filter(psal_qc == 1) %>% 
@@ -116,7 +115,7 @@ prof_file <- system.file(
 )
 
 argo_read_prof_levels(prof_file)
-#> # A tibble: 70 x 17
+#> # A tibble: 70 × 17
 #>    N_LEVELS N_PROF   PRES PRES_QC PRES_ADJUSTED PRES_ADJUSTED_QC
 #>  *    <int>  <int>  <dbl> <chr>           <dbl> <chr>           
 #>  1        1      1   9.80 1                9.80 1               
@@ -129,7 +128,7 @@ argo_read_prof_levels(prof_file)
 #>  8        8      1  80.3  1               80.3  1               
 #>  9        9      1  90.2  1               90.2  1               
 #> 10       10      1 100.   1              100.   1               
-#> # ... with 60 more rows, and 11 more variables: PRES_ADJUSTED_ERROR <dbl>,
+#> # … with 60 more rows, and 11 more variables: PRES_ADJUSTED_ERROR <dbl>,
 #> #   TEMP <dbl>, TEMP_QC <chr>, TEMP_ADJUSTED <dbl>, TEMP_ADJUSTED_QC <chr>,
 #> #   TEMP_ADJUSTED_ERROR <dbl>, PSAL <dbl>, PSAL_QC <chr>, PSAL_ADJUSTED <dbl>,
 #> #   PSAL_ADJUSTED_QC <chr>, PSAL_ADJUSTED_ERROR <dbl>
