+++ README.Rmd
@@ -30,7 +30,7 @@ argo_global_prof()
 [![Codecov test coverage](https://codecov.io/gh/ArgoCanada/argodata/branch/master/graph/badge.svg)](https://codecov.io/gh/ArgoCanada/argodata?branch=master)
 <!-- badges: end -->
 
-The goal of argodata is to provide a data frame-based interface to data generated by the [Argo floats program](https://argo.ucsd.edu). This package provides low-level access to all variables contained within Argo NetCDF files; for a higher-level interface with built-in visualization and quality control, see the [argoFloats package](https://github.com/ArgoCanada/argoFloats).
+The goal of argodata is to provide a data frame-based interface to data generated by the [Argo floats program](https://argo.ucsd.edu) ([doi:10.17882/42182](https://doi.org/10.17882/42182)). This package provides low-level access to all variables contained within Argo NetCDF files; for a higher-level interface with built-in visualization and quality control, see the [argoFloats package](https://github.com/ArgoCanada/argoFloats).
 
 ## Installation
 
@@ -41,7 +41,7 @@ You can install the development version from [GitHub](https://github.com/) using
 remotes::install_github("ArgoCanada/argodata")
 ```
 
-The argodata package downloads files from the [FTP and HTTPS mirrors](http://www.argodatamgt.org/Access-to-data/Access-via-FTP-or-HTTPS-on-GDAC), caches them, and loads them into R. You can set the mirror using `argo_set_mirror()` and the cache directory using `argo_set_cache_dir()`:
+The argodata package downloads files from the [FTP and HTTPS mirrors](http://www.argodatamgt.org/Access-to-data/Argo-GDAC-ftp-and-https-servers), caches them, and loads them into R. You can set the mirror using `argo_set_mirror()` and the cache directory using `argo_set_cache_dir()`:
 
 ```{r, eval = FALSE}
 argo_set_mirror("https://data-argo.ifremer.fr/")
