### 2021-10-05 80c3384
try to add another dep to pkgdown
