### 2021-10-05 fa3d45c
document return value of `argo_global_*()` functions (#37)
