### 2021-10-05 ea672b8
see if running in macos helps
