+++ .github/workflows/pkgdown.yaml
@@ -9,18 +9,13 @@ name: pkgdown
 
 jobs:
   pkgdown:
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
       - uses: actions/checkout@v2
-
       - uses: r-lib/actions/setup-pandoc@v1
-
       - uses: r-lib/actions/setup-r@v1
-        with:
-          use-public-rspm: true
-
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
           extra-packages: pkgdown, rmarkdown, lubridate, ggplot2, ggspatial, lwgeom, s2
