### 2021-10-06 #3659 @paleolimbot
Thanks @KimBaldry for the initial review! I believe I've worked through the first round of issues but feel free to let me know if I haven't!
