### 2021-10-08 #3659 @catsch
@kthyng If I raised some issues for example regarding the statement of need should I check the lines in the checklist or should I wait before checking an answer from the author ?
