### 2021-10-09 #3659 @KimBaldry
> **Submitting author:** @paleolimbot ([Dewey Dunnington](http://orcid.org/0000-0002-9415-4582)) **Repository:** https://github.com/ArgoCanada/argodata **Version:** 0.0.0.9000 **Editor:** @kthyng **Reviewer:** @KimBaldry, @catsch **Archive:** Pending
>
> **⚠️ JOSS reduced service mode ⚠️**
>
> Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
>
> ## Status
> [![status](https://camo.githubusercontent.com/62efa8db56988b8e4bff904b11a190b1580a160d582c400a91e14b391a480e3d/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f39393066616431323132663736346665326538316362646664306436323037662f7374617475732e737667)](https://joss.theoj.org/papers/990fad1212f764fe2e81cbdfd0d6207f)
>
> Status badge code:
>
> ```
> ```
>
> **Reviewers and authors**:
>
> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/ArgoCanada/argodata) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
>
> ## Reviewer instructions & questions
> @KimBaldry & @catsch, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
>
> 1. Make sure you're logged in to your GitHub account
> 2. Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
>
> The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let @kthyng know.
>
> ✨ **Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest** ✨
>
> ## Review checklist for @KimBaldry
> ✨ Important: Please do not use the _Convert to issue_ functionality when working through this checklist, instead, please open any new issues associated with your review [in the software repository associated with the submission](https://github.com/ArgoCanada/argodata). ✨
>
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/ArgoCanada/argodata)?
> * [x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [x]  **Contribution and authorship:** Has the submitting author (@paleolimbot) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [x]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * []  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [x]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [x]  **A statement of need:** Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
> * [x]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [x]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

@kthyng I am having trouble editing the checklist directly. My initial review covered everything but functionality. I note a lack of community guidelines for contributing to the software and submitting issues (@paleolimbot an extra paragraph in the read me will suffice). I will test functionality in the next 1-2 weeks and finalise my review.
