### 2021-10-13 d319534
add argo_filter_parameter_data_mode() (#43)
