+++ NAMESPACE
@@ -11,6 +11,7 @@ export(argo_filter_date)
 export(argo_filter_direction)
 export(argo_filter_float)
 export(argo_filter_parameter)
+export(argo_filter_parameter_data_mode)
 export(argo_filter_radius)
 export(argo_filter_rect)
 export(argo_filter_updated)
