### 2021-10-13 5f0a435
fix reading of parameter_data_mode column (make sure leading/trailing whitespace is preserved) (#43)
