### 2021-10-13 edbfda3
expand on the role of BGC variables in the need for argodata (#44)
