### 2021-10-13 e78c1e2
clarify that `argo_download()` is usually not called manually since it is called automatically by the load functions (#45)
