### 2021-10-13 60c62c5
clarify the four file types as NetCDF file formats and reference the Argo User's Manual (#46)
