+++ inst/joss/paper.bib
@@ -204,8 +204,12 @@
   number = {1}
 }
 
-
-
+@techreport{argo-user-manual,
+  year = "2021",
+  title = "Argo user's manual V3.41",
+  author = {Argo Data Management Team},
+  doi = {https://doi.org/10.13155/29825}
+}
 
 @Manual{r-rnoaa,
   title = {rnoaa: 'NOAA' Weather Data from R},
