### 2021-10-13 e4eceb0
add BGC reference to summary (#44), better wording around `argo_download()` (#45)
