+++ inst/joss/paper.bib
@@ -208,7 +208,7 @@
   year = "2021",
   title = "Argo user's manual V3.41",
   author = {Argo Data Management Team},
-  doi = {https://doi.org/10.13155/29825}
+  doi = {10.13155/29825}
 }
 
 @Manual{r-rnoaa,
