### 2021-10-13 #3659 @paleolimbot
Thank you for the initial round of comments! I've left replies to the issues on the argodata repo and look forward to any additional comments you may have! I've done my best to address the initial comments on the text...let me know what else is needed.
