### 2021-11-22 #3659 @catsch
yes it is

--
Catherine Schmechtig

0681827411

LEFE-CYBER Database
http://www.obs-vlfr.fr/proof

BGC Argo
http://biogeochemical-argo.org/

https://github.com/catsch

OSU Ecce Terra
Sorbonne Université - CNRS - UMS 3455
4, Place Jussieu
75 252 Paris Cedex 05
Tour 45 - couloir 45-55
2eme etage - 207

De: "Kristen Thyng" ***@***.***>
À: "openjournals" ***@***.***>
Cc: "Catherine Schmechtig" ***@***.***>, "Mention" ***@***.***>
Envoyé: Lundi 22 Novembre 2021 17:04:11
Objet: Re: [openjournals/joss-reviews] [REVIEW]: argodata: An R interface to oceanographic data from the International Argo Program (#3659)

[ https://github.com/catsch | @catsch ] Oops! It looks like you're right! Sorry for the extraneous ping. Can I verify then that the submission is acceptable from your perspective?
