### 2021-12-07 #3659 @kthyng
Ok excellent! We can proceed with the next steps then.

@paleolimbot Would you like to tag a new version of your software? I suggest jumping up to a non-beta sounding version number with this advancement. Let me know what version to associate with this JOSS submission.

Also, please archive your code with a place like Zenodo, and then report here your doi. Be sure to update the metadata in the Zenodo archive so that the title and author list exactly match your JOSS paper.
