### 2021-12-07 #3659 @kthyng
@paleolimbot Paper comments:
* Line 31: Is something missing after "chlorophyll"?
* in the references, there is an "r" that isn't capitalized, and "welcome to the tidyverse" is there twice. Please correct these and look through all references to see if there are any other errors.
