+++ inst/joss/paper.md
@@ -44,7 +44,7 @@ This paper describes ``argodata``, an R package that makes it easier to work wit
 
 The ocean is highly variable in both space and time and mapping this variability at appropriate scales is a key factor in many scientific studies. Oceanographic data have direct applications that range from the analysis of near-bottom ecosystems to air-sea interactions.  More broadly, ocean measurements are needed to constrain the models that scientists use to understand the evolving state of the ocean and to make predictions about its future, particularly as a component of the global climate system.
 
-The International Argo Program [@argo] deploys and collects data from several thousand devices that are programmed to drift with and move vertically through the ocean. Sensors measure electrical conductivity, temperature, pressure, and other quantities along this vertical path yielding "profiles" that are uploaded via satellite to globally distributed data assembly centres [@roemmich_argo_2001; @roemmich_argo_2009-2]. Since 1997, the International Argo Program has collected over 2.4 million profiles from around the globe and expanded its original array of sensors to measure biogeochemicical variables such as pH, chloryphyll-α, dissolved oxygen, nitrate, and many others.
+The International Argo Program [@argo] deploys and collects data from several thousand devices that are programmed to drift with and move vertically through the ocean. Sensors measure electrical conductivity, temperature, pressure, and other quantities along this vertical path yielding "profiles" that are uploaded via satellite to globally distributed data assembly centres [@roemmich_argo_2001; @roemmich_argo_2009-2]. Since 1997, the International Argo Program has collected over 2.4 million profiles from around the globe and expanded its original array of sensors to measure biogeochemicical variables such as pH, chloryphyll-a, dissolved oxygen, nitrate, and many others.
 
 Although the NetCDF data files provided by Argo data servers contain metadata that describe their contents, we identified a number of barriers to data access. These included (1) reading and decoding the index files to locate files of interest, (2) downloading and potentially caching large numbers of small NetCDF files, (3) reading the NetCDF files into a form where the data contained within can be visualized and analyzed, and (4) dealing efficiently with potentially large Argo data sets. In particular, the incorporation of biogeochemical variables in Argo NetCDF files introduced additional complexity such that a novice- to average-level programmer may have difficulty extracting and manipulating data from many profiles. Whereas a variety of applications have been created to address some of these barriers, the ``argodata`` package is our attempt to overcome these barriers for the novice- to average-level programmer who may not be familiar with oceanographic conventions for storing data.
 
@@ -89,7 +89,7 @@ To locate files of interest on the Argo mirror, index files for profile, traject
 ```
 
 ```
-## # A tibble: 2,524,364 x 8
+## # A tibble: 2,556,123 x 8
 ##    file     date                latitude longitude ocean profiler_type
 ##    <chr>    <dttm>                 <dbl>     <dbl> <chr>         <dbl>
 ##  1 aoml/13~ 1997-07-29 20:03:00    0.267     -16.0 A               845
@@ -102,7 +102,7 @@ To locate files of interest on the Argo mirror, index files for profile, traject
 ##  8 aoml/13~ 1997-10-14 18:39:35    1.76      -21.6 A               845
 ##  9 aoml/13~ 1997-10-25 19:32:34    1.80      -21.8 A               845
 ## 10 aoml/13~ 1997-11-05 18:51:42    1.64      -21.4 A               845
-## # ... with 2,524,354 more rows, and 2 more variables:
[…]
