+++ inst/joss/paper.bib
@@ -212,7 +212,7 @@
 }
 
 @Manual{r-rnoaa,
-  title = {rnoaa: 'NOAA' Weather Data from R},
+  title = {rnoaa: 'NOAA' Weather Data from {R}},
   author = {Scott Chamberlain},
   year = {2021},
   note = {R package version 1.3.2},
@@ -249,18 +249,6 @@
   abstract = {Wickham et al., (2019). Welcome to the Tidyverse. Journal of Open Source Software, 4(43), 1686, https://doi.org/10.21105/joss.01686},
   langid = {english},
   number = {43}
-}
-
-@article{r-tidyverse,
-  title = {Welcome to the tidyverse},
-  doi = {10.21105/joss.01686},
-  url = {https://doi.org/10.21105/joss.01686},
-  year = {2019},
-  publisher = {The Open Journal},
-  volume = {4},
-  number = {43},
-  pages = {1686},
-  journal = {Journal of Open Source Software}
 }
 
 @Manual{r-tidync,
