+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: argodata
 Title: Locate, Download, and Read Argo Ocean Float Data
-Version: 0.0.0.9000
+Version: 0.1.0
 Authors@R: c(
     person(given = "Dewey",
            family = "Dunnington",
