+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.7.3
-pkgdown: 1.6.1
+pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
-  articles/virtual-mooring: virtual-mooring.html
-last_built: 2021-10-13T19:25Z
+  virtual-mooring: virtual-mooring.html
+last_built: 2021-12-10T13:54Z
 urls:
   reference: https://argocanada.github.io/argodata/reference
   article: https://argocanada.github.io/argodata/articles
