+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   virtual-mooring: virtual-mooring.html
-last_built: 2021-12-10T13:54Z
+last_built: 2021-12-10T14:01Z
 urls:
   reference: https://argocanada.github.io/argodata/reference
   article: https://argocanada.github.io/argodata/articles
