### 2021-12-10 #3659 @paleolimbot
Ok! I think I've hit all the bits from the comments above:

- The version is 0.1.0
- The Zenodo DOI is 10.5281/zenodo.5772495
- Fixed the minor edits noted in the above comment

Let me know if anything else is required!
